🛠️ refactor: consolidate fragmented HTTPS fetchers into a centralized, hardened `httpGet` utility
